+++ psnotify/psnotify_linux.go
@@ -9,6 +9,8 @@ import (
 	"encoding/binary"
 	"os"
 	"syscall"
+
+	"github.com/elastic/gosigar/sys"
 )
 
 const (
@@ -30,7 +32,7 @@ const (
 )
 
 var (
-	byteOrder = binary.LittleEndian
+	byteOrder = sys.GetEndian()
 )
 
 // linux/connector.h: struct cb_id