+++ CHANGELOG.md
@@ -5,7 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
-- Added support for huge TLB pages on Linux #97
+- Added support for huge TLB pages on Linux #97  
+- Added support for big endian platform  
 
 ### Fixed